Remove the subfolder tooltips preference
There is no ui for it and as far as I recall no one ever asked for one.
Let's just always enable them. One less preference to port to gsettings.
The code in ringschart is still in place, but if this ever needs to be
brought back it should become a property.