rework code for excluded locations
Instead of keeping a list attached to the Baobab.Application and having
a method there for checking it, attach a list to the scanner object and
create it at the start of the scan.

This prevents the scanner from calling back into the application class
from the worker thread.

Also, switch to using a hashtable.  We have at least 4 items in it (plus
the ones that the user added) so this should already be slightly faster
than a linear scan.