Bubble up scanner errors.
For now this is very strict: we abort the scan on every error, which
means that scan filesystem will never work unless running as root. We
probably need to be more permissive and simply mark directories where
we do not have permissions and warn that the results may be not
accurate.