Ringchart: always use the correct state when getting colors
The ringchart was calling get_background_color on the toplevel context
with a fixed state of NORMAL without saving the context first. If the
toplevel however had additional state flags set, such as BACKDROP, this
could trigger an invalidation and cause a redraw, resulting in a redraw
loop.

Instead of saving/restoring the context, this patch uses the current
state of the context, which is the right thing to do in general anyway
and does not require saving/restoring the context unless the context is
actually modified.

https://bugzilla.gnome.org/show_bug.cgi?id=754703