Add ability to scan as administrator
When we encounter an inaccessible subdirectory, we try the GIO URI "admin:///path/to/directory".
If the user fails to authenticate, we just go on with the regular scan.

Since the "admin://" scheme is used only for subdirectories that can't be read as
a regular user, it shouldn't be *too* much slower, hopefully.