Port to GTK4+libadwaita
This commit ports baobab to GTK and makes use of libadwaita
widgets and styling.

Things depending on Events have been reworked to use
shortcuts and eventcontrollers, like drag and drop or
the various shortcuts for the treeview.

There have also been tweaks to the styling made to fit
in better with the new Adwaita look and feel.