Opt-in to libhandy dark style preference
GNOME 42 will have a system-wide dark style preference,
libadwaita and libhandy now provide API for supporting it.
libadwaita is planned to be part of the GNOME platform for
the next release, it would be good to have our applications
follow it.

This commit adds support for the new dark style via
libhandy.