window: Remove even if the file does not exist
Previously, when the user tried to move a file to the trash
when the file had already been removed, the action would fail,
and the entry would stay in the list even though it was not
on the file system anymore.

So do not consider GLib.IOError.NOT_FOUND a failure, and still
go through with the removal.