scanner: Harden and simplify the display name
This directly stores the display name in the model as it will be
rendered, ensuring it's always a valid non-null string. The operation
was already half there when creating the model's items anyway and was
already generating all the strings, but not storing the relevant one.
This allows removing functions creating the actually displayable string,
simplifying the code, which will come handy when reworking the model and
widgets.

Signed-off-by: Adrien Plazas <aplazas@gnome.org>