Disable column reordering of FolderDisplay
Currently, it is possible to reorder columns of the `BaobabFolderDisplay`
widget using drag and drop. However, there are some issues with this
feature. The new column order is applied only to the first row and not to
the others that are part of the separate `treeview` widget. Another
problem	is that dragging the "Folder" column sometimes significantly
increases the window width. Let's disable the column reordering until
those issues are resolved.

Related: https://issues.redhat.com/browse/RHEL-40874
Related: https://issues.redhat.com/browse/RHEL-40875