fix: Popup menus acting on wrong items
Both the column list view and chart views have bugs where due to the
mechanism used to retrieve the item to be actioned the wrong item can be
used depending on where the mouse has moved. This easily results in data
loss by moving the wrong file or folder to trash.

Most of the baobab-window.vala changes are Maximiliano's - @msandova.

Closes #157
Closes #183